dev, release 버전 db 분리

+++ microservices/book-service/src/main/java/com/dto/ReviewResponseDto.java
@@ -11,7 +11,7 @@ public class ReviewResponseDto {
 
     private String identifier;
 
-    private Integer avgReviewRating;
+    private Double avgReviewRating;
 
     private Integer reviewCount;
